Minimize amount of code split between browser and node
