+++ lib/util/node-utils-mock.js
@@ -1,0 +1,4 @@
+// In browser environments, this function is a no-op.
+export function nodeFilenameToStream( filename ) {
+	return filename;
+}
